webhook/requirements-dev.txt
astroid==2.4.2
attrs==19.3.0
autopep8==1.5.3
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
fastapi==0.59.0
filelock==3.0.12
flake8==3.8.3
get==0.0.0
google-api-core==1.21.0
google-auth==1.19.1
google-cloud-core==1.3.0
google-cloud-datastore==1.13.1
google-cloud-logging==1.15.0
google-cloud-ndb==1.4.1
googleapis-common-protos==1.52.0
grpcio==1.30.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.7.15
httpcore==0.9.1
httptools==0.1.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pkg-resources==0.0.0
pluggy==0.13.1
post==0.0.0
protobuf==3.12.2
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pydantic==1.6
pyflakes==2.2.0
pylint==2.5.3
pyparsing==2.4.7
pytest==5.4.3
pytonic==0.0.1.dev1
pytz==2020.1
query-string==0.0.0
redis==3.5.3
request==0.0.0
requests==2.24.0
rfc3986==1.4.0
rsa==4.6
six==1.15.0
sniffio==1.1.0
starlette==0.13.4
toml==0.10.1
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
wcwidth==0.2.5
websockets==8.1
wrapt==1.12.1
